Move TypeDecl classes to separate header
