+++ CMakeLists.txt
@@ -12,6 +12,7 @@ add_definitions(${LLVM_DEFINITIONS})
 # Include and lib directories
 include_directories(SYSTEM ${LLVM_INCLUDE_DIRS}) 
 include_directories(${CMAKE_SOURCE_DIR}/runtime/include)
+include_directories(${CMAKE_SOURCE_DIR}/compiler/include)
 link_directories(${CMAKE_SOURCE_DIR}/runtime/target/release)
 
 # Add runtime project
